Update on "[WIP][xpu][feature] Support get_device_tflops and get_gpu_dram_gbps for XPU."
This PR add the basic support of `get_device_tflops` and `get_gpu_dram_gbps` for XPU and enabled the corresponding test case.

cc voznesenskym penguinwu EikanWang jgong5 Guobing-Chen XiaobingSuper zhuhaozhe blzheng wenzhe-nrv jiayisunx ipiszy kadeng muchulee8 amjames chauhang aakhundov coconutruben jataylo Lucaskabela

[ghstack-poisoned]

+++ torch/_inductor/utils.py
@@ -2779,7 +2779,7 @@ def get_device_tflops(dtype: torch.dtype) -> float:
 
 
 @functools.cache
-def get_gpu_dram_gbps() -> int:
+def get_gpu_dram_gbps() -> float:
     dram_bw_gbs = datasheet_dram_bw_gbs()
     if dram_bw_gbs is not None:
         return dram_bw_gbs
